Now checking for @deprecated location correctness and moved around the implicit types
