+++ src/main/java/graphql/Directives.java
@@ -1,6 +1,10 @@
 package graphql;
 
 
+import graphql.language.Description;
+import graphql.language.DirectiveDefinition;
+import graphql.language.InputValueDefinition;
+import graphql.language.StringValue;
 import graphql.schema.GraphQLDirective;
 
 import static graphql.Scalars.GraphQLBoolean;
@@ -11,6 +15,9 @@ import static graphql.introspection.Introspection.DirectiveLocation.FIELD_DEFINI
 import static graphql.introspection.Introspection.DirectiveLocation.FRAGMENT_SPREAD;
 import static graphql.introspection.Introspection.DirectiveLocation.INLINE_FRAGMENT;
 import static graphql.introspection.Introspection.DirectiveLocation.SCALAR;
+import static graphql.language.DirectiveLocation.newDirectiveLocation;
+import static graphql.language.NonNullType.newNonNullType;
+import static graphql.language.TypeName.newTypeName;
 import static graphql.schema.GraphQLArgument.newArgument;
 import static graphql.schema.GraphQLNonNull.nonNull;
 
@@ -19,6 +26,38 @@ import static graphql.schema.GraphQLNonNull.nonNull;
  */
 @PublicApi
 public class Directives {
+
+    public static final String NO_LONGER_SUPPORTED = "No longer supported";
+    public static final DirectiveDefinition DEPRECATED_DIRECTIVE_DEFINITION;
+    public static final DirectiveDefinition SPECIFIED_BY_DIRECTIVE_DEFINITION;
+
+    static {
+        DEPRECATED_DIRECTIVE_DEFINITION = DirectiveDefinition.newDirectiveDefinition()
+                .name(Directives.DeprecatedDirective.getName())
+                .directiveLocation(newDirectiveLocation().name(FIELD_DEFINITION.name()).build())
+                .directiveLocation(newDirectiveLocation().name(ENUM_VALUE.name()).build())
+                .description(createDescription("Marks the field or enum value as deprecated"))
+                .inputValueDefinition(
+                        InputValueDefinition.newInputValueDefinition()
+                                .name("reason")
+                                .description(createDescription("The reason for the deprecation"))
+                                .type(newTypeName().name("String").build())
+                                .defaultValue(StringValue.newStringValue().value(NO_LONGER_SUPPORTED).build())
+                                .build())
+                .build();
+
+        SPECIFIED_BY_DIRECTIVE_DEFINITION = DirectiveDefinition.newDirectiveDefinition()
+                .name(Directives.SpecifiedByDirective.getName())
+                .directiveLocation(newDirectiveLocation().name(SCALAR.name()).build())
+                .description(createDescription("Exposes a URL that specifies the behaviour of this scalar."))
+                .inputValueDefinition(
+                        InputValueDefinition.newInputValueDefinition()
+                                .name("url")
+                                .description(createDescription("The URL that specifies the behaviour of this scalar."))
+                                .type(newNonNullType(newTypeName().name("String").build()).build())
+                                .build())
+                .build();
+    }
 
     public static final GraphQLDirective IncludeDirective = GraphQLDirective.newDirective()
             .name("include")
@@ -43,7 +82,7 @@ public class Directives {
 
     /**
      * The "deprecated" directive is special and is always available in a graphql schema
-     *
+     * <p>
      * See https://graphql.github.io/graphql-spec/June2018/#sec--deprecated
      */
     public static final GraphQLDirective DeprecatedDirective = GraphQLDirective.newDirective()
@@ -52,9 +91,10 @@ public class Directives {
             .argument(newArgument()
                     .name("reason")
                     .type(GraphQLString)
-                    .defaultValue("No longer supported")
+                    .defaultValue(NO_LONGER_SUPPORTED)
                     .description("The reason for the deprecation"))
             .validLocations(FIELD_DEFINITION, ENUM_VALUE)
+            .definition(DEPRECATED_DIRECTIVE_DEFINITION)
             .build();
 
     /**
@@ -68,6 +108,11 @@ public class Directives {
                     .type(nonNull(GraphQLString))
                     .description("The URL that specifies the behaviour of this scalar."))
             .validLocations(SCALAR)
+            .definition(SPECIFIED_BY_DIRECTIVE_DEFINITION)
             .build();
+
+    private static Description createDescription(String s) {
+        return new Description(s, null, false);
+    }
 
 }
