+++ src/main/java/graphql/schema/idl/SchemaTypeDirectivesChecker.java
@@ -126,10 +126,6 @@ class SchemaTypeDirectivesChecker {
 
     private void checkDirectives(DirectiveLocation expectedLocation, List<GraphQLError> errors, TypeDefinitionRegistry typeRegistry, Node<?> element, String elementName, List<Directive> directives) {
         directives.forEach(directive -> {
-            // we have special code for the deprecation directive in SchemaTypeChecker.checkDeprecatedDirective
-            if (Directives.DeprecatedDirective.getName().equals(directive.getName())) {
-                return;
-            }
             Optional<DirectiveDefinition> directiveDefinition = typeRegistry.getDirectiveDefinition(directive.getName());
             if (!directiveDefinition.isPresent()) {
                 errors.add(new DirectiveUndeclaredError(element, elementName, directive.getName()));
